# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Era Inventory API
  description: |
    # Era Inventory Management API
    
    A comprehensive REST API for managing inventory items, sites, vendors, and projects with multi-tenant organization isolation.
    
    ## Features
    - **Multi-tenant Architecture**: Organization-based data isolation using Row Level Security (RLS)
    - **Role-based Access Control**: Fine-grained permissions (org_admin, project_admin)
    - **JWT Authentication**: Secure token-based authentication
    - **Comprehensive CRUD Operations**: Full lifecycle management for all entities
    - **Advanced Search & Filtering**: Text search with pagination and sorting
    - **Error Handling**: Detailed error codes and messages for all scenarios
    
    ## Authentication
    All API endpoints (except health checks) require JWT authentication via the `Authorization: Bearer <token>` header.
    
    ### JWT Token Structure
    ```json
    {
      "sub": "1",           // User ID
      "org_id": 1,          // Organization ID
      "roles": ["org_admin"], // User roles
      "iss": "era-inventory",
      "aud": "era-inventory-api",
      "exp": 1640995200,
      "iat": 1640908800,
      "nbf": 1640908800
    }
    ```
    
    ### Role Permissions
    - **org_admin**: Full access to all operations within the organization
    - **project_admin**: Can create/update items, read-only access to other entities
    
    ## Error Handling
...
    
    ## Rate Limiting
...
    
  version: 1.0.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /items:
    get:
      summary: List Inventory Items
      description: |
        Retrieve a paginated list of inventory items for the authenticated user's organization.
        
        ### Features
        - **Pagination**: Use `limit` and `offset` parameters
        - **Search**: Use `q` parameter to search by name or asset_tag
        - **Sorting**: Use `sort` parameter (e.g., `name`, `-created_at`)
        - **Organization Isolation**: Only returns items from user's organization
        
        ### Search Examples
        - `?q=switch` - Find items with "switch" in name or asset_tag
        - `?sort=name` - Sort by name ascending
        - `?sort=-created_at` - Sort by creation date descending
        - `?limit=10&offset=20` - Get items 21-30
      tags: [Items]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of inventory items
          headers:
            X-Total-Count:
              description: Total number of items matching the filter
              schema:
                type: integer
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Item'
              examples:
                items_list:
                  summary: List of items
                  value:
                    data:
                      - id: 1
                        asset_tag: "SW-001"
                        name: "Access Switch"
                        manufacturer: "Cisco"
                        model: "C2960X"
                        device_type: "Switch"
                        site: "Main Office"
                        installed_at: "2023-01-15T10:30:00Z"
                        warranty_end: "2026-01-15T00:00:00Z"
                        notes: "24-port gigabit switch"
                        created_at: "2023-01-10T09:00:00Z"
                        updated_at: "2023-01-15T10:30:00Z"
                      - id: 2
                        asset_tag: "RT-001"
                        name: "Core Router"
                        manufacturer: "Juniper"
                        model: "EX4300"
                        device_type: "Router"
                        site: "Data Center"
                        installed_at: "2023-01-20T14:00:00Z"
                        warranty_end: "2026-01-20T00:00:00Z"
                        notes: null
                        created_at: "2023-01-18T11:00:00Z"
                        updated_at: "2023-01-20T14:00:00Z"
                    page:
                      limit: 50
                      offset: 0
                      total: 2
                empty_list:
                  summary: Empty list
                  value:
                    data: []
                    page:
                      limit: 50
                      offset: 0
                      total: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create Inventory Item
      description: |
        Create a new inventory item in the authenticated user's organization.
        
        ### Required Permissions
        - `org_admin` or `project_admin` role
        
        ### Validation Rules
        - `asset_tag` must be unique within the organization
        - `name` is required and cannot be empty
        - Date fields must be in ISO 8601 format
        
        ### Business Rules
        - Item is automatically associated with the user's organization
        - `created_at` and `updated_at` are automatically set
      tags: [Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
            examples:
              basic_item:
                summary: Basic item creation
                value:
                  asset_tag: "SW-002"
                  name: "Distribution Switch"
                  manufacturer: "Cisco"
                  model: "C3850"
                  device_type: "Switch"
                  site: "Branch Office"
              full_item:
                summary: Complete item with all fields
                value:
                  asset_tag: "SRV-001"
                  name: "Web Server"
                  manufacturer: "Dell"
                  model: "PowerEdge R740"
                  device_type: "Server"
                  site: "Data Center"
                  installed_at: "2023-02-01T09:00:00Z"
                  warranty_end: "2026-02-01T00:00:00Z"
                  notes: "Primary web application server"
      responses:
        '201':
          description: Item created successfully
          headers:
            Location:
              description: URL of the created item
              schema:
                type: string
                format: uri
                example: "/items/123"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                created_item:
                  summary: Successfully created item
                  value:
                    id: 3
                    asset_tag: "SW-002"
                    name: "Distribution Switch"
                    manufacturer: "Cisco"
                    model: "C3850"
                    device_type: "Switch"
                    site: "Branch Office"
                    installed_at: null
                    warranty_end: null
                    notes: null
                    created_at: "2023-02-15T14:30:00Z"
                    updated_at: "2023-02-15T14:30:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_required:
                  summary: Missing required fields
                  value:
                    error: "Validation failed: name is required"
                    code: "VALIDATION_ERROR"
                invalid_date:
                  summary: Invalid date format
                  value:
                    error: "Invalid date format for installed_at"
                    code: "INVALID_DATE_FORMAT"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Asset tag conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate_asset_tag:
                  summary: Asset tag already exists
                  value:
                    error: "Asset tag 'SW-002' already exists"
                    code: "DUPLICATE_ASSET_TAG"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /items/{id}:
    parameters:
      - $ref: '#/components/parameters/ItemIdParam'

    get:
      summary: Get Inventory Item
      description: |
        Retrieve a specific inventory item by ID.
        
        ### Organization Isolation
        Only returns items from the authenticated user's organization.
      tags: [Items]
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                item_details:
                  summary: Complete item details
                  value:
                    id: 1
                    asset_tag: "SW-001"
                    name: "Access Switch"
                    manufacturer: "Cisco"
                    model: "C2960X"
                    device_type: "Switch"
                    site: "Main Office"
                    installed_at: "2023-01-15T10:30:00Z"
                    warranty_end: "2026-01-15T00:00:00Z"
                    notes: "24-port gigabit switch"
                    created_at: "2023-01-10T09:00:00Z"
                    updated_at: "2023-01-15T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update Inventory Item
      description: |
        Update an existing inventory item.
        
        ### Required Permissions
        - `org_admin` or `project_admin` role
        
        ### Validation Rules
        - `asset_tag` must remain unique within the organization
        - Cannot update items from other organizations
        - All fields are optional in the request body
        
        ### Business Rules
        - `updated_at` is automatically updated
        - Partial updates are supported
      tags: [Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
            examples:
              partial_update:
                summary: Update only some fields
                value:
                  name: "Updated Switch Name"
                  notes: "Updated notes"
              full_update:
                summary: Update all fields
                value:
                  asset_tag: "SW-001-UPD"
                  name: "Updated Access Switch"
                  manufacturer: "Cisco"
                  model: "C2960X-48TD-L"
                  device_type: "Switch"
                  site: "Updated Office"
                  installed_at: "2023-01-20T10:30:00Z"
                  warranty_end: "2027-01-20T00:00:00Z"
                  notes: "Updated 48-port gigabit switch"
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                updated_item:
                  summary: Successfully updated item
                  value:
                    id: 1
                    asset_tag: "SW-001-UPD"
                    name: "Updated Access Switch"
                    manufacturer: "Cisco"
                    model: "C2960X-48TD-L"
                    device_type: "Switch"
                    site: "Updated Office"
                    installed_at: "2023-01-20T10:30:00Z"
                    warranty_end: "2027-01-20T00:00:00Z"
                    notes: "Updated 48-port gigabit switch"
                    created_at: "2023-01-10T09:00:00Z"
                    updated_at: "2023-02-15T16:45:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Asset tag conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate_asset_tag:
                  summary: Asset tag already exists
                  value:
                    error: "Asset tag 'SW-001-UPD' already exists"
                    code: "DUPLICATE_ASSET_TAG"
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete Inventory Item
      description: |
        Delete an inventory item.
        
        ### Required Permissions
        - `org_admin` role only
        
        ### Business Rules
        - Cannot delete items from other organizations
        - Deletion is permanent and cannot be undone
        - Associated data may be affected
      tags: [Items]
      responses:
        '204':
          description: Item deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permissions (org_admin required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_permissions:
                  summary: Only org_admin can delete items
                  value:
                    error: "Insufficient permissions"
                    code: "INSUFFICIENT_PERMISSIONS"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Sites Endpoints
  /sites:
    get:
      summary: List Sites
      description: |
        Retrieve a paginated list of sites for the authenticated user's organization.
        
        ### Features
        - **Pagination**: Use `limit` and `offset` parameters
        - **Search**: Use `q` parameter to search by site name
        - **Sorting**: Use `sort` parameter
        - **Organization Isolation**: Only returns sites from user's organization
      tags: [Sites]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of sites
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Site'
              examples:
                sites_list:
                  summary: List of sites
                  value:
                    data:
                      - id: 1
                        name: "Main Office"
                        location: "123 Business St, City, State 12345"
                        notes: "Primary headquarters location"
                        created_at: "2023-01-01T00:00:00Z"
                        updated_at: "2023-01-01T00:00:00Z"
                      - id: 2
                        name: "Data Center"
                        location: "456 Tech Blvd, Tech City, State 67890"
                        notes: null
                        created_at: "2023-01-05T00:00:00Z"
                        updated_at: "2023-01-05T00:00:00Z"
                    page:
                      limit: 50
                      offset: 0
                      total: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create Site
      description: |
        Create a new site in the authenticated user's organization.
        
        ### Required Permissions
        - `org_admin` role
        
        ### Validation Rules
        - `name` is required and cannot be empty
        - `location` and `notes` are optional
      tags: [Sites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteInput'
            examples:
              basic_site:
                summary: Basic site creation
                value:
                  name: "Branch Office"
                  location: "789 Branch Ave, Branch City, State 11111"
              minimal_site:
                summary: Site with name only
                value:
                  name: "Remote Location"
      responses:
        '201':
          description: Site created successfully
          headers:
            Location:
              description: URL of the created site
              schema:
                type: string
                format: uri
                example: "/sites/123"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sites/{id}:
    parameters:
      - $ref: '#/components/parameters/SiteIdParam'

    get:
      summary: Get Site
      description: |
        Retrieve a specific site by ID.
        
        ### Organization Isolation
        Only returns sites from the authenticated user's organization.
      tags: [Sites]
      responses:
        '200':
          description: Site details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update Site
      description: |
        Update an existing site.
        
        ### Required Permissions
        - `org_admin` role
      tags: [Sites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteInput'
      responses:
        '200':
          description: Site updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete Site
      description: |
        Delete a site.
        
        ### Required Permissions
        - `org_admin` role
      tags: [Sites]
      responses:
        '204':
          description: Site deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Vendors Endpoints
  /vendors:
    get:
      summary: List Vendors
      description: |
        Retrieve a paginated list of vendors for the authenticated user's organization.
      tags: [Vendors]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of vendors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Vendor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create Vendor
      description: |
        Create a new vendor in the authenticated user's organization.
        
        ### Required Permissions
        - `org_admin` role
      tags: [Vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInput'
      responses:
        '201':
          description: Vendor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vendors/{id}:
    parameters:
      - $ref: '#/components/parameters/VendorIdParam'

    get:
      summary: Get Vendor
      description: Retrieve a specific vendor by ID.
      tags: [Vendors]
      responses:
        '200':
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update Vendor
      description: |
        Update an existing vendor.
        
        ### Required Permissions
        - `org_admin` role
      tags: [Vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInput'
      responses:
        '200':
          description: Vendor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete Vendor
      description: |
        Delete a vendor.
        
        ### Required Permissions
        - `org_admin` role
      tags: [Vendors]
      responses:
        '204':
          description: Vendor deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Projects Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Include in the Authorization header as: `Bearer <token>`
        
        ### Token Generation
        Use the `/auth/login` endpoint or the `jwtgen` utility to generate tokens.
        
        ### Token Claims
        - `sub`: User ID
        - `org_id`: Organization ID
        - `roles`: Array of user roles
        - `exp`: Token expiration timestamp
        
        ### Example
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  parameters:
...
    ItemIdParam:
      name: id
      in: path
      required: true
      description: Unique identifier of the inventory item
      schema:
        type: integer
        minimum: 1
      example: 123
...